Extracted gs-shared into gluestick
