--- packages/gluestick-shared/.babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [
-    "react",
-    "es2015", 
-    "stage-0"
-  ]
-}
-
