--- packages/gluestick-shared/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-npm-debug.log
-.DS_Store
-build
-.idea/
