--- packages/gluestick-shared/.npmignore
@@ -1,2 +1,0 @@
-**/__mocks__/**
-**/__tests__/**
