--- packages/gluestick-shared/README.md
@@ -1,5 +1,0 @@
-# GlueStick Shared
-[![npm](https://img.shields.io/npm/v/gluestick-shared.svg)](https://www.npmjs.com/package/gluestick-shared)
-[![Build Status](https://travis-ci.org/TrueCar/gluestick-shared.svg?branch=master)](https://travis-ci.org/TrueCar/gluestick-shared)
-
-This contains the shared code in the GlueStick CLI and the apps that it generates.
