--- packages/gluestick-shared/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "gluestick-shared",
-  "version": "1.0.0-rc.7",
-  "description": "This contains the shared code in the GlueStick CLI and the apps that it generates.",
-  "main": "./build/index.js",
-  "scripts": {
-    "test": "jest",
-    "test-coverage": "jest --coverage",
-    "clean": "rimraf build",
-    "build": "./node_modules/.bin/babel src --out-dir build",
-    "prepublish": "npm run clean && npm run build"
-  },
-  "files": [
-    "build"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/TrueCar/gluestick-shared.git"
-  },
-  "author": "Todd Williams",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/TrueCar/gluestick-shared/issues"
-  },
-  "homepage": "https://github.com/TrueCar/gluestick-shared#readme",
-  "dependencies": {
-    "react-router-scroll": "0.3.2",
-    "react-side-effect": "1.0.2",
-    "redux-immutable-state-invariant": "1.2.2",
-    "redux-thunk": "2.0.1",
-    "axios": "0.15.3"
-  },
-  "jest": {
-    "notify": true,
-    "testRegex": ".*/__tests__/.*\\.test\\.js$"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.8.0",
-    "babel-loader": "6.4.0",
-    "babel-polyfill": "^6.8.0",
-    "babel-preset-es2015": "6.22.0",
-    "babel-preset-react": "6.23.0",
-    "babel-preset-stage-0": "6.22.0",
-    "jest": "19.0.0",
-    "radium": ">=0.17.1",
-    "react": ">=15.1.0",
-    "react-dom": ">=15.1.0",
-    "react-redux": ">=4.4.5",
-    "react-router": ">=2.4.1",
-    "redux": ">=3.5.2",
-    "rimraf": "^2.5.0"
-  }
-}
