+++ packages/gluestick/package.json
@@ -53,7 +53,6 @@
     "extract-text-webpack-plugin": "2.0.0-beta.5",
     "fs-extra": "2.0.0",
     "gluestick-generators": "^1.0.0-rc.7",
-    "gluestick-shared": "^1.0.0-rc.7",
     "handlebars": "4.0.6",
     "http-proxy-middleware": "0.17.3",
     "inquirer": "3.0.1",
