+++ packages/gluestick/shared/lib/prepareRoutesWithTransitionHooks.js
